+++ src/azure-cli/azure/cli/command_modules/storage/__init__.py
@@ -23,13 +23,13 @@ class StorageCommandsLoader(AzCommandsLoader):
 
     def load_command_table(self, args):
         super(StorageCommandsLoader, self).load_command_table(args)
-        from azure.cli.command_modules.storage.commands_v2 import load_command_table
+        from azure.cli.command_modules.storage.commands import load_command_table
         load_command_table(self, args)
         return self.command_table
 
     def load_arguments(self, command):
         super(StorageCommandsLoader, self).load_arguments(command)
-        from azure.cli.command_modules.storage._params_v2 import load_arguments
+        from azure.cli.command_modules.storage._params import load_arguments
         load_arguments(self, command)
 
 
@@ -46,13 +46,13 @@ class AzureStackStorageCommandsLoader(AzCommandsLoader):
 
     def load_command_table(self, args):
         super(AzureStackStorageCommandsLoader, self).load_command_table(args)
-        from azure.cli.command_modules.storage.commands import load_command_table
+        from azure.cli.command_modules.storage.commands_azure_stack import load_command_table
         load_command_table(self, args)
         return self.command_table
 
     def load_arguments(self, command):
         super(AzureStackStorageCommandsLoader, self).load_arguments(command)
-        from azure.cli.command_modules.storage._params import load_arguments
+        from azure.cli.command_modules.storage._params_azure_stack import load_arguments
         load_arguments(self, command)
 
 
@@ -283,7 +283,7 @@ def _merge_new_exception_handler(kwargs, handler):
 
 
 def get_command_loader(cli_ctx):
-    if cli_ctx.cloud.profile != 'latest':
+    if cli_ctx.cloud.profile.lower() != 'latest':
         return AzureStackStorageCommandsLoader
 
     return StorageCommandsLoader
