+++ src/azure-cli/azure/cli/command_modules/storage/_params_azure_stack.py
