+++ src/azure-cli/azure/cli/command_modules/storage/commands_azure_stack.py
